StreamlitMarkdown element
Wraps <ReactMarkdown> and supplies our plugins and renderers
